api/requirements.txt
fastapi
uvicorn[standard]
pandas
pymongo
pydantic
google-generativeai
python-dotenv
python-multipart